requirements.txt
datasets
chromadb
open-clip-torch
google-generativeai
streamlit
requests
langchain
langchain-community
openai
tiktoken